Move InMemorySketchesStore
