+++ java/system-test/system-test-dsl/pom.xml
@@ -116,6 +116,13 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
+            <artifactId>ingest-runner</artifactId>
+            <version>${project.parent.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>sleeper</groupId>
             <artifactId>ingest-batcher-core</artifactId>
             <version>${project.parent.version}</version>
             <type>test-jar</type>
