+++ java/system-test/system-test-dsl/src/test/java/sleeper/systemtest/dsl/testutil/drivers/InMemoryPartitionSplittingDriver.java
@@ -20,6 +20,7 @@ import org.slf4j.LoggerFactory;
 
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.statestore.StateStore;
+import sleeper.ingest.runner.testutils.InMemorySketchesStore;
 import sleeper.splitter.core.find.FindPartitionsToSplit;
 import sleeper.splitter.core.find.FindPartitionsToSplit.JobSender;
 import sleeper.splitter.core.split.SplitPartition;
